pyproject.toml
[tool.poetry]
name = "vbdist"
version = "0.1.0"
description = ""
authors = ["Himmelsw4nderer <71812274+Himmelsw4nderer@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
sounddevice = "^0.4.6"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"
pynacl = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
